#315 天啊，累死宝宝了啊

+++ src/main/java/com/taobao/rigel/rap/account/dao/AccountDao.java
@@ -194,6 +194,6 @@ public interface AccountDao {
      * @param teamId
      * @return
      */
-    List<Integer> getUserIdList(int teamId);
+    List<Integer> getUserIdList(long teamId);
 
 }
