+++ src/main/java/com/taobao/rigel/rap/account/dao/impl/AccountDaoImpl.java
@@ -115,10 +115,10 @@ public class AccountDaoImpl extends HibernateDaoSupport implements AccountDao {
         currentSession().update(user);
     }
 
-    public List<Integer> getUserIdList(int teamId) {
+    public List<Integer> getUserIdList(long teamId) {
         String sql = "SELECT user_id FROM tb_corporation_and_user WHERE corporation_id = :teamId";
         Query query = currentSession().createSQLQuery(sql);
-        query.setInteger("teamId", teamId);
+        query.setLong("teamId", teamId);
         return query.list();
     }
