+++ src/main/java/com/taobao/rigel/rap/project/bo/Project.java
@@ -343,7 +343,7 @@ public class Project implements java.io.Serializable, EntityInterface {
         return checkInListOrdered;
     }
 
-    public Module findModule(int moduleId) {
+    public Module findModule(long moduleId) {
         for (Module i : getModuleList()) {
             if (i.getId() == moduleId)
                 return i;
@@ -351,7 +351,7 @@ public class Project implements java.io.Serializable, EntityInterface {
         return null;
     }
 
-    public Page findPage(int pageId) {
+    public Page findPage(long pageId) {
         for (Module module : getModuleList()) {
             for (Page page : module.getPageList()) {
                 if (page.getId() == pageId)
@@ -373,7 +373,7 @@ public class Project implements java.io.Serializable, EntityInterface {
         return null;
     }
 
-    public Parameter findParameter(int parameterId, boolean isRequestType) {
+    public Parameter findParameter(long parameterId, boolean isRequestType) {
         for (Module module : getModuleList()) {
             for (Page page : module.getPageList()) {
                 for (Action action : page.getActionList()) {
@@ -389,7 +389,7 @@ public class Project implements java.io.Serializable, EntityInterface {
         return null;
     }
 
-    public Parameter findChildParameter(int parameterId) {
+    public Parameter findChildParameter(long parameterId) {
         for (Module module : getModuleList()) {
             for (Page page : module.getPageList()) {
                 for (Action action : page.getActionList()) {
@@ -421,7 +421,7 @@ public class Project implements java.io.Serializable, EntityInterface {
      * @param id
      * @return return the object found, other wise return null
      */
-    private Parameter findParameterRecursively(Parameter p, int id) {
+    private Parameter findParameterRecursively(Parameter p, long id) {
         Iterator<Parameter> iterator = p.getParameterList().iterator();
         while (iterator.hasNext()) {
             Parameter parameter = iterator.next();
@@ -436,7 +436,7 @@ public class Project implements java.io.Serializable, EntityInterface {
         return null;
     }
 
-    public void removeModule(int id, Session session) {
+    public void removeModule(long id, Session session) {
         Module module = findModule(id);
         if (module != null && moduleList != null) {
             moduleList.remove(module);
@@ -444,7 +444,7 @@ public class Project implements java.io.Serializable, EntityInterface {
         }
     }
 
-    public void removePage(int id, Session session) {
+    public void removePage(long id, Session session) {
         Page page = findPage(id);
         if (page != null && page.getModule() != null && page.getModule().getPageList() != null) {
             page.getModule().getPageList().remove(page);
@@ -452,7 +452,7 @@ public class Project implements java.io.Serializable, EntityInterface {
         }
     }
 
-    public void removeAction(int id, Session session) {
+    public void removeAction(long id, Session session) {
         Action action = findAction(id);
         if (action == null) return;
         Set<Page> pageList = action.getPageList();
@@ -464,7 +464,7 @@ public class Project implements java.io.Serializable, EntityInterface {
         }
     }
 
-    public void removeParameter(int id, Session session) {
+    public void removeParameter(long id, Session session) {
         Parameter parameter = findParameter(id, true);
 
         // if parameter == null, it maybe response parameter
